+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Backtrace Android Release Notes
 
+## Version 3.5.0 - 14.09.2021
+- Added support for native crash reporting in NDK 16b
+- Bug fixes and expanded supported NDK versions for client side unwinding
+
 ## Version 3.4.0 - 07.09.2021
 - Added support for NDK 22